pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "snowflake-agentic-platform"
version = "1.0.0"
description = "Zero-to-One Agentic Data Platform - From use case description + raw data to deployed AI application"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.10"
authors = [
    {name = "Snowflake", email = "support@snowflake.com"}
]
keywords = ["snowflake", "ai", "agents", "langgraph", "cortex", "ml"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "snowflake-connector-python>=3.6.0",
    "snowflake-snowpark-python>=1.14.0",
    "snowflake-ml-python>=1.5.0",
    "langgraph>=0.2.0",
    "langchain>=0.2.0",
    "langchain-core>=0.2.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "typer>=0.12.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "mypy>=1.8.0",
    "ruff>=0.3.0",
    "black>=24.0.0",
    "pre-commit>=3.6.0",
]
observability = [
    "langsmith>=0.1.0",
    "langfuse>=2.0.0",
]
all = [
    "snowflake-agentic-platform[dev,observability]",
]

[project.scripts]
agentic-platform = "src.cli.main:app"

[project.urls]
Homepage = "https://github.com/snowflake/snowflake-agentic-platform"
Documentation = "https://github.com/snowflake/snowflake-agentic-platform/docs"
Repository = "https://github.com/snowflake/snowflake-agentic-platform"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
markers = [
    "unit: Unit tests (no external dependencies)",
    "integration: Integration tests (requires Snowflake connection)",
    "e2e: End-to-end tests (full pipeline)",
    "slow: Slow running tests",
]

[tool.mypy]
python_version = "3.11"
warn_return_any = false
warn_unused_configs = true
ignore_missing_imports = true
check_untyped_defs = false
disallow_untyped_defs = false
no_implicit_optional = false
strict = false
disable_error_code = ["union-attr", "attr-defined", "call-arg", "arg-type", "return-value", "assignment", "operator", "typeddict-item", "no-any-return", "import-untyped", "override", "type-var", "index", "var-annotated", "valid-type"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501", "N806"]  # N806: Allow uppercase variable names (sklearn convention X, X_train, etc.)

[tool.ruff.lint.per-file-ignores]
"src/app/backend/main.py" = ["E402"]  # Module level import not at top (FastAPI pattern)

[tool.black]
line-length = 100
target-version = ["py311"]
